Refactor paf-le-chien and fix several issues about ocaml-h2
From an unikernel opam-mirror and a simple http client, it seems that we
have a leak of file-descriptors. This last commit do 2 main things:
- find a common trunck betwen ocaml-h2 and httpaf
- refactor alpn and give a type witness instead of multiple adts
- pass a shutdown function which call the underlying shutdown of the
  runtime ({Httpaf,H2}.{Server,Client}_connection.shutdown

The last point should be well documented, the main issue is: shutdown
can be out of the scope of the server handler. Depending on the usage,
the user can probably shutdown the HTTP (http/1.1 or h2) connection or
let the runtime (http/1.1 or h2) to properly close the connection. Due
to the bidirectional communication over HTTP, it's hard to figure out
without the user usage when we should shutdown so we prefer to give
such opportunity to the end-user.

For instance, a Body.flush ; Body.close ; shutdown ; should works but we
did not get such case when the scheduling of the runtime is hard to
formalize...

This version was tested with http-lwt-client and a unikernel as a server
which can be used as a proxy with HTTPS (so, the CONNECT method seems to
work.

For this last case, we put the ability to avoid the runtime to really
close the underlying connection. Indeed, for the proxy, we expect that
the request handler finishes and the underlying fiber (the main loop)
finishes too. However, the last one finishes with a Flow.close and, in
the context of the proxy, the communication must persist. So we signals
the underlying flow to fake the next close, finish the request and reset
the underlying flow to properly close if we call Flow.close.

Voilà, voilà...